% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_shp.R
\name{write_shp}
\alias{write_shp}
\title{Write geodata to a shapefile}
\usage{
write_shp(geodata, dsn = getwd(), layer = deparse(substitute(geodata)), ...,
  verbose = TRUE)
}
\arguments{
\item{geodata}{(sf or sp object)}

\item{dsn}{(character) generally, a path on disk; see \code{\link[sf:st_write]{sf::st_write()}}}

\item{layer}{(character) layer name; see \code{\link[sf:st_write]{sf::st_write()}}}

\item{...}{further arguments to \code{\link[sf:write_sf]{sf::write_sf()}}}
}
\value{
\code{geodata}, invisibly (to allow for chaining)
}
\description{
Write geodata to a shapefile
}
\examples{
require(ARB)
dsn <- tempdir()
write_shp(ARB_CoAbDis, dsn = dsn, layer = "CoAbDis", delete_layer = TRUE)

}
