% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip_only.R
\name{unzip_only}
\alias{unzip_only}
\title{Unzip only files matching a pattern}
\usage{
unzip_only(zip_file, pattern = NULL, junkpaths = FALSE, ...,
  verbose = getOption("verbose"))
}
\arguments{
\item{zip_file}{(character) path to ZIP file}

\item{pattern}{(character) see \link[utils:glob2rx]{utils::glob2rx} and \link[stringr:regex]{stringr::regex}}

\item{junkpaths}{(logical) if \code{TRUE}, use only the basename of the stored filepath (discard subdirectory names)}

\item{...}{further arguments to \link[utils:unzip]{utils::unzip}}

\item{verbose}{(logical)}
}
\description{
Unzip only files matching a pattern
}
\examples{
zip_file <- "/Users/dholstius/GitHub/BAAQMD/ARB/tmp/ca_co_ab_dis.zip"
exdir <- unzip_only(zip_file, pattern = "CoAbDis", verbose = TRUE)
show(exdir)

}
